Moved cal info loader to utils and improved handling
